# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Petshop Api",
        "description": "L5 Swagger Api documentation for Buckhill Challenge",
        "contact": {
            "email": "yusufbenaiah@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Demo API/V1 Server for challenge"
        }
    ],
    "paths": {
...
        "api/v1/admin/user-listing/{uuid}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "operationId": "35b97c012c7697cb47f17034173efdc2",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID parameter",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "address",
                                    "phone_number"
                                ],
                                "properties": {
                                    "first_name": {
                                        "description": "User First Name",
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "description": "User First Name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "User email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string"
                                    },
                                    "avatar": {
                                        "description": "User password",
                                        "type": "string"
                                    },
                                    "address": {
                                        "description": "User address",
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "description": "User phone number",
                                        "type": "string"
                                    },
                                    "is_marketing": {
                                        "description": "User marketing preferences",
                                        "type": "string",
                                        "enum": [
                                            "0",
                                            "1"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "operationId": "f9ac5d1ef466271e8566ef0a542270cd",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID parameter",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "api/v1/admin/create": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "operationId": "408f115e227105cdb5b87e51111c84bd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "address",
                                    "phone_number"
                                ],
                                "properties": {
                                    "first_name": {
                                        "description": "User First Name",
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "description": "User First Name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "User email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string"
                                    },
                                    "avatar": {
                                        "description": "User password",
                                        "type": "string"
                                    },
                                    "address": {
                                        "description": "User address",
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "description": "User phone number",
                                        "type": "string"
                                    },
                                    "is_marketing": {
                                        "description": "User marketing preferences",
                                        "type": "string",
                                        "enum": [
                                            "0",
                                            "1"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    }
}